Tests/LibWeb/Text/expected/geometry/getClientRects-wrapped-span-with-border-padding.txt

wrap: 3
wrap[0]: 10,7 43x26
wrap[1]: 5,27 40x26
wrap[2]: 5,47 43x26
